// lib/types/branch.ts
export interface BranchListItem {
  branchListId: number;
  branchCode: string;
  branchName: string;
  dateOpened?: string;
  regionCode?: string;
  regionName?: string;
  districtCode?: string;
  districtName?: string;
  coreBank?: string;
  inputter?: string;
  authoriser?: string;
  transDate?: string;
  transTime?: string;
  auditorCode?: string;
  auditDateTime?: string;
  branchMnemonic?: string;
  createdAt?: string;
  updatedAt?: string;
}